The charm of algorithm
About how learning algorithm, the following, the individual briefly talk about three points (learning attitude algorithm = interest + + code + focus on practical application) :
Interest in it. To learn any thing, it must first have the interest. Algorithm is no exception. If you really not interested to algorithm, also don't need to worry. After all, his taste.
Attitude, and try again. If you think that because the algorithm is applied in the actual work is not much, or standard library encapsulates everything. Or think that the most basic of quick sort don't have to go to understand, so this article can sweep through the.
Practice. Reading, such as data structure of the teaching material, be sure to take all kinds of most basic data structures, arrays, strings, stack, stack, queue, tree and graph make TongTongTouTou (in fact, look at me before finishing the 100 Microsoft interview questions, there is also a great help, because the interview questions are associated with data structures and algorithms, and that's what I was doing over), the second is days like friends say, basic iterative method, the exhaustive search method, recursive method, recursive method, greedy method, partition method, dynamic programming, backtracking, and so on are to understand to get through. Can then be read programming of mission, introduction to algorithms (a lot of people say introduction to algorithms don't understand, but is it a lot of mathematical proof, I don't really understand. Painting can be more specific on the paper, such as red and black tree related operations and code), plus the scrutiny - thinking repeatedly, repeatedly research, finally coding -- attention and combined with the practical application, coded (write code implements a algorithm is more useful than research through an algorithm), and to really understand a data structure and algorithms, must be combined with the actual project. Only above, without it.
Below, I recommend you a link to the learning algorithm
http://blog.csdn.net/v_july_v/article/details/6543438